Refactored and consolidated codebase
