--- VideoPlayer/Models/ControllerSettings.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace VideoPlayer.Models
-{
-    public class ControllerSettings
-    {
-        public int ControllerSelectButton { get; set; }
-        public int ControllerBackButton { get; set; }
-    }
-}
